refactor (useNAvigation): rename to navigation provider
