+++ src/popup/hooks/index.ts
@@ -2,5 +2,4 @@ export * from './useData'
 export * from './useFavicons'
 export * from './useBookmarkStatus'
 export * from './useSearch'
-export * from './useNavigate'
 export * from './useColorScheme'
