--- src/popup/hooks/useNavigate.ts
@@ -1,6 +1,0 @@
-import { useContext } from "react";
-import { RouterContext } from "../providers/RouterProvider";
-
-export const useNavigate = () => {
-  return useContext(RouterContext);
-}
