+++ src/popup/screens/Overview/OverviewHeader.tsx
@@ -1,9 +1,9 @@
 import { VFC } from 'react';
 import { Separator } from '../../components/Separator';
-import { useNavigate } from '../../hooks';
 import { OverviewHeaderLatestItems } from './OverviewHeaderLatestItems';
 import { OverviewHeaderUpperItem } from './OverviewHeaderUpperItem';
 import { overviewHeaderContainer } from './OverviewHeader.css';
+import { useNavigate } from '../../providers';
 
 type OverviewHeaderProps = {
   openTabs: number;
